# spec/swagger.json
{
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Welcome to the Ory Kratos HTTP API documentation!",
    "title": "Ory Kratos",
    "version": "latest"
  },
  "basePath": "/",
  "paths": {
...
    "/self-service/registration": {
      "post": {
        "description": "Use this endpoint to complete a registration flow by sending an identity's traits and password. This endpoint\nbehaves differently for API and browser flows.\n\nAPI flows expect `application/json` to be sent in the body and respond with\nHTTP 200 and a application/json body with the created identity success - if the session hook is configured the\n`session` and `session_token` will also be included;\nHTTP 303 redirect to a fresh registration flow if the original flow expired with the appropriate error messages set;\nHTTP 400 on form validation errors.\n\nBrowser flows expect a Content-Type of `application/x-www-form-urlencoded` or `application/json` to be sent in the body and respond with\na HTTP 303 redirect to the post/after registration URL or the `return_to` value if it was set and if the registration succeeded;\na HTTP 303 redirect to the registration UI URL with the flow ID containing the validation errors otherwise.\n\nBrowser flows with an accept header of `application/json` will not redirect but instead respond with\nHTTP 200 and a application/json body with the signed in identity and a `Set-Cookie` header on success;\nHTTP 303 redirect to a fresh login flow if the original flow expired with the appropriate error messages set;\nHTTP 400 on form validation errors.\n\nIf this endpoint is called with `Accept: application/json` in the header, the response contains the flow without a redirect. In the\ncase of an error, the `error.id` of the JSON response body can be one of:\n\n`session_already_available`: The user is already signed in.\n`security_csrf_violation`: Unable to fetch the flow because a CSRF violation occurred.\n`security_identity_mismatch`: The requested `?return_to` address is not allowed to be used. Adjust this in the configuration!\n`browser_location_change_required`: Usually sent when an AJAX request indicates that the browser needs to open a specific URL.\nMost likely used in Social Sign In flows.\n\nMore information can be found at [Ory Kratos User Login and User Registration Documentation](https://www.ory.sh/docs/next/kratos/self-service/flows/user-login-user-registration).",
        "consumes": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "v0alpha2"
        ],
        "summary": "Submit a Registration Flow",
        "operationId": "submitSelfServiceRegistrationFlow",
        "parameters": [
          {
            "type": "string",
            "description": "The Registration Flow ID\n\nThe value for this parameter comes from `flow` URL Query parameter sent to your\napplication (e.g. `/registration?flow=abcde`).",
            "name": "flow",
            "in": "query",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/submitSelfServiceRegistrationFlowBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfulSelfServiceRegistrationWithoutBrowser",
            "schema": {
              "$ref": "#/definitions/successfulSelfServiceRegistrationWithoutBrowser"
            }
          },
          "303": {
            "$ref": "#/responses/emptyResponse"
          },
          "400": {
            "description": "selfServiceRegistrationFlow",
            "schema": {
              "$ref": "#/definitions/selfServiceRegistrationFlow"
            }
          },
          "422": {
            "description": "selfServiceBrowserLocationChangeRequiredError",
            "schema": {
              "$ref": "#/definitions/selfServiceBrowserLocationChangeRequiredError"
            }
          },
          "500": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          }
        }
      }
    },
    "/self-service/registration/api": {
      "get": {
        "description": "This endpoint initiates a registration flow for API clients such as mobile devices, smart TVs, and so on.\n\nIf a valid provided session cookie or session token is provided, a 400 Bad Request error\nwill be returned unless the URL query parameter `?refresh=true` is set.\n\nTo fetch an existing registration flow call `/self-service/registration/flows?flow=\u003cflow_id\u003e`.\n\nYou MUST NOT use this endpoint in client-side (Single Page Apps, ReactJS, AngularJS) nor server-side (Java Server\nPages, NodeJS, PHP, Golang, ...) browser applications. Using this endpoint in these applications will make\nyou vulnerable to a variety of CSRF attacks.\n\nIn the case of an error, the `error.id` of the JSON response body can be one of:\n\n`session_already_available`: The user is already signed in.\n`security_csrf_violation`: Unable to fetch the flow because a CSRF violation occurred.\n\nThis endpoint MUST ONLY be used in scenarios such as native mobile apps (React Native, Objective C, Swift, Java, ...).\n\nMore information can be found at [Ory Kratos User Login and User Registration Documentation](https://www.ory.sh/docs/next/kratos/self-service/flows/user-login-user-registration).",
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "v0alpha2"
        ],
        "summary": "Initialize Registration Flow for APIs, Services, Apps, ...",
        "operationId": "initializeSelfServiceRegistrationFlowWithoutBrowser",
        "responses": {
          "200": {
            "description": "selfServiceRegistrationFlow",
            "schema": {
              "$ref": "#/definitions/selfServiceRegistrationFlow"
            }
          },
          "400": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          },
          "500": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          }
        }
      }
    },
    "/self-service/registration/browser": {
      "get": {
        "description": "This endpoint initializes a browser-based user registration flow. This endpoint will set the appropriate\ncookies and anti-CSRF measures required for browser-based flows.\n\n:::info\n\nThis endpoint is EXPERIMENTAL and subject to potential breaking changes in the future.\n\n:::\n\nIf this endpoint is opened as a link in the browser, it will be redirected to\n`selfservice.flows.registration.ui_url` with the flow ID set as the query parameter `?flow=`. If a valid user session\nexists already, the browser will be redirected to `urls.default_redirect_url`.\n\nIf this endpoint is called via an AJAX request, the response contains the flow without a redirect. In the\ncase of an error, the `error.id` of the JSON response body can be one of:\n\n`session_already_available`: The user is already signed in.\n`security_csrf_violation`: Unable to fetch the flow because a CSRF violation occurred.\n`security_identity_mismatch`: The requested `?return_to` address is not allowed to be used. Adjust this in the configuration!\n\nIf this endpoint is called via an AJAX request, the response contains the registration flow without a redirect.\n\nThis endpoint is NOT INTENDED for clients that do not have a browser (Chrome, Firefox, ...) as cookies are needed.\n\nMore information can be found at [Ory Kratos User Login and User Registration Documentation](https://www.ory.sh/docs/next/kratos/self-service/flows/user-login-user-registration).",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "v0alpha2"
        ],
        "summary": "Initialize Registration Flow for Browsers",
        "operationId": "initializeSelfServiceRegistrationFlowForBrowsers",
        "parameters": [
          {
            "type": "string",
            "description": "The URL to return the browser to after the flow was completed.",
            "name": "return_to",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "selfServiceRegistrationFlow",
            "schema": {
              "$ref": "#/definitions/selfServiceRegistrationFlow"
            }
          },
          "303": {
            "$ref": "#/responses/emptyResponse"
          },
          "500": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          }
        }
      }
    },
    "/self-service/registration/flows": {
      "get": {
        "description": "This endpoint returns a registration flow's context with, for example, error details and other information.\n\nBrowser flows expect the anti-CSRF cookie to be included in the request's HTTP Cookie Header.\nFor AJAX requests you must ensure that cookies are included in the request or requests will fail.\n\nIf you use the browser-flow for server-side apps, the services need to run on a common top-level-domain\nand you need to forward the incoming HTTP Cookie header to this endpoint:\n\n```js\npseudo-code example\nrouter.get('/registration', async function (req, res) {\nconst flow = await client.getSelfServiceRegistrationFlow(req.header('cookie'), req.query['flow'])\n\nres.render('registration', flow)\n})\n```\n\nThis request may fail due to several reasons. The `error.id` can be one of:\n\n`session_already_available`: The user is already signed in.\n`self_service_flow_expired`: The flow is expired and you should request a new one.\n\nMore information can be found at [Ory Kratos User Login and User Registration Documentation](https://www.ory.sh/docs/next/kratos/self-service/flows/user-login-user-registration).",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "v0alpha2"
        ],
        "summary": "Get Registration Flow",
        "operationId": "getSelfServiceRegistrationFlow",
        "parameters": [
          {
            "type": "string",
            "description": "The Registration Flow ID\n\nThe value for this parameter comes from `flow` URL Query parameter sent to your\napplication (e.g. `/registration?flow=abcde`).",
            "name": "id",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "HTTP Cookies\n\nWhen using the SDK on the server side you must include the HTTP Cookie Header\noriginally sent to your HTTP handler here.",
            "name": "cookie",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "selfServiceRegistrationFlow",
            "schema": {
              "$ref": "#/definitions/selfServiceRegistrationFlow"
            }
          },
          "403": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          },
          "404": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          },
          "410": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          },
          "500": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UUID": {"type": "string", "format": "uuid4"},
...
    "identityCredentialsType": {
      "description": "and so on.",
      "type": "string",
      "title": "CredentialsType  represents several different credential types, like password credentials, passwordless credentials,"
    },
...
    "jsonError": {
      "description": "The standard Ory JSON API error format.",
      "type": "object",
      "title": "JSON API Error Response",
      "required": [
        "error"
      ],
      "properties": {
        "error": {
          "$ref": "#/definitions/genericError"
        }
      }
    },
...
    "selfServiceBrowserLocationChangeRequiredError": {
      "type": "object",
      "title": "Is sent when a flow requires a browser to change its location.",
      "required": [
        "message"
      ],
      "properties": {
        "code": {
          "description": "The status code",
          "type": "integer",
          "format": "int64",
          "example": 404
        },
        "debug": {
          "description": "Debug information\n\nThis field is often not exposed to protect against leaking\nsensitive information.",
          "type": "string",
          "example": "SQL field \"foo\" is not a bool."
        },
        "details": {
          "description": "Further error details",
          "type": "object",
          "additionalProperties": true
        },
        "id": {
          "description": "The error ID\n\nUseful when trying to identify various errors in application logic.",
          "type": "string"
        },
        "message": {
          "description": "Error message\n\nThe error's message.",
          "type": "string",
          "example": "The resource could not be found"
        },
        "reason": {
          "description": "A human-readable reason for the error",
          "type": "string",
          "example": "User with ID 1234 does not exist."
        },
        "redirect_browser_to": {
          "description": "Since when the flow has expired",
          "type": "string"
        },
        "request": {
          "description": "The request ID\n\nThe request ID is often exposed internally in order to trace\nerrors across service architectures. This is often a UUID.",
          "type": "string",
          "example": "d7ef54b1-ec15-46e6-bccb-524b82c035e6"
        },
        "status": {
          "description": "The status description",
          "type": "string",
          "example": "Not Found"
        }
      }
    },
...
    "selfServiceFlowType": {
      "description": "The flow type can either be `api` or `browser`.",
      "type": "string",
      "title": "Type is the flow type."
    },
...
    "selfServiceRegistrationFlow": {
      "type": "object",
      "required": [
        "id",
        "expires_at",
        "issued_at",
        "request_url",
        "ui"
      ],
      "properties": {
        "active": {
          "$ref": "#/definitions/identityCredentialsType"
        },
        "expires_at": {
          "description": "ExpiresAt is the time (UTC) when the flow expires. If the user still wishes to log in,\na new flow has to be initiated.",
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "$ref": "#/definitions/UUID"
        },
        "issued_at": {
          "description": "IssuedAt is the time (UTC) when the flow occurred.",
          "type": "string",
          "format": "date-time"
        },
        "request_url": {
          "description": "RequestURL is the initial URL that was requested from Ory Kratos. It can be used\nto forward information contained in the URL's path or query for example.",
          "type": "string"
        },
        "return_to": {
          "description": "ReturnTo contains the requested return_to URL.",
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/selfServiceFlowType"
        },
        "ui": {
          "$ref": "#/definitions/uiContainer"
        }
      }
    },
...
    "session": {
      "description": "A Session",
      "type": "object",
      "required": [
        "id",
        "identity"
      ],
      "properties": {
        "active": {
          "description": "Active state. If false the session is no longer active.",
          "type": "boolean"
        },
        "authenticated_at": {
          "description": "The Session Authentication Timestamp\n\nWhen this session was authenticated at. If multi-factor authentication was used this\nis the time when the last factor was authenticated (e.g. the TOTP code challenge was completed).",
          "type": "string",
          "format": "date-time"
        },
        "authentication_methods": {
          "$ref": "#/definitions/sessionAuthenticationMethods"
        },
        "authenticator_assurance_level": {
          "$ref": "#/definitions/authenticatorAssuranceLevel"
        },
        "expires_at": {
          "description": "The Session Expiry\n\nWhen this session expires at.",
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "$ref": "#/definitions/UUID"
        },
        "identity": {
          "$ref": "#/definitions/identity"
        },
        "issued_at": {
          "description": "The Session Issuance Timestamp\n\nWhen this session was issued at. Usually equal or close to `authenticated_at`.",
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "submitSelfServiceRegistrationFlowBody": {
      "type": "object"
    },
...
    "successfulSelfServiceRegistrationWithoutBrowser": {
      "description": "The Response for Registration Flows via API",
      "type": "object",
      "required": [
        "identity"
      ],
      "properties": {
        "identity": {
          "$ref": "#/definitions/identity"
        },
        "session": {
          "$ref": "#/definitions/session"
        },
        "session_token": {
          "description": "The Session Token\n\nThis field is only set when the session hook is configured as a post-registration hook.\n\nA session token is equivalent to a session cookie, but it can be sent in the HTTP Authorization\nHeader:\n\nAuthorization: bearer ${session-token}\n\nThe session token is only issued for API flows, not for Browser flows!",
          "type": "string"
        }
      }
    },
...
  },
  "responses": {
    "emptyResponse": {
      "description": "Empty responses are sent when, for example, resources are deleted. The HTTP status code for empty responses is typically 201."
    }
  },
  "securityDefinitions": {
    "oryAccessToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}